dev: add: commands based folder structure
