+++ src/express/commands/destroy.js
@@ -1,0 +1,10 @@
+const shell = require("shelljs");
+
+export async function terraformDestroy() {
+    console.log("📍Executing terraform destroy...")
+    shell.exec(`terraform destroy -auto-approve`, {
+        env: {
+            ...process.env,
+        }
+    });
+}
