+++ src/express/commands/init.js
@@ -1,0 +1,10 @@
+const shell = require("shelljs");
+
+export async function terraformInit() {
+    console.log("📍Executing terraform init...")
+    shell.exec(`terraform init`, {
+        env: {
+            ...process.env,
+        }
+    });
+}
